Projects/04-AI_Story_Creator/requirements.txt
streamlit>=1.20.0
openai>=1.0.0
python-dotenv>=1.0.0
Pillow>=9.0.0
requests>=2.28.0
huggingface_hub>=0.16.0 